refactor(attributes): extract ATTR_HELP into its own module
Pure move ahead of porting the attribute detail widget to Preact. No
behaviour change: the map and its types are unchanged, only exported.
